+++ changelog.md
@@ -10,6 +10,9 @@
 * Fast foods min speed to eat has been reduced (0.75s -> 0.6s)
 * Smokers now require soul sand/soil to craft
 * Twins chance has been moved to data packs (livestock_data)
+* Hoes Definitions
+  * Renamed from Hoes Stats
+  * Removed hoes cooldown from Hoes Definition
 * Changed, moved and renamed many config options
   * Client
     * World border has its own feature now